Crate governance_os_runtime

This module hosts the runtime for the governance OS, it bundles the other pallets from the governance OS project. It aims at providing the following:

a demonstration runtime for anybody willing to test the pallets

a reference runtime for anybody trying to use the pallets in their own runtime

Modules

api

Structs

CoinVotingParameters

GenesisConfig

NativeCurrencyId

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PlcrVotingParameters

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionKeys

Enums

Call

Event

OriginCaller

RuntimeVotingParameters
An enum to wrap the different voting parameters for all the runtime voting implementations.

RuntimeVotingSystemId
An enum to differentiate between the different runtime voting implementations. The voting router will use it to select the right parameters and vote data.

Constants

VERSION
Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version information used to identify this runtime when compiled natively. Typically used when building the Executor in our node.

Type Definitions

Address
The address format for describing accounts.

AllModules
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem
All pallets included in the runtime as a nested tuple of types.

Aura

AuraConfig

AuraId
An Aura authority identifier using S/R 25519 as its crypto.

Block
Block type as expected by this runtime.

Bylaws

BylawsConfig

CoinVoting

Compat

ConvictionVoting

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

GrandpaId
Identity of a Grandpa authority.

Header
Block header type as expected by this runtime.

NativeCurrency
The system’s native currency, typically used to pay for fees.

Organizations

OrganizationsConfig

PlcrVoting

RandomnessCollectiveFlip

SignedExtra
The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

Tokens

TokensConfig

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.