Crate governance_os_runtime

This module hosts the runtime for the governance OS, it bundles the other pallets from the governance OS project. It aims at providing the following:

a demonstration runtime for anybody willing to test the pallets

a reference runtime for anybody trying to use the pallets in their own runtime

Modules

api

Structs

GenesisConfig

NativeCurrencyId

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionKeys

Enums

Call

Event

OriginCaller

Constants

VERSION
Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
The version information used to identify this runtime when compiled natively. Typically used when building the Executor in our node.

Type Definitions

Address
The address format for describing accounts.

Aura

AuraConfig

AuraId
An Aura authority identifier using S/R 25519 as its crypto.

Block
Block type as expected by this runtime.

Bylaws

BylawsConfig

Compat

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

GrandpaId
Identity of a Grandpa authority.

Header
Block header type as expected by this runtime.

NativeCurrency
The system's native currency, typically used to pay for fees.

Organizations

OrganizationsConfig

RandomnessCollectiveFlip

SignedExtra
The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

Tokens

TokensConfig

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.